Refactored Adams add Manager to guild into editGuildView
Might be some functionality that needs tweaking
