+++ src/museumtimetracking/gui/views/root/activeGuilds/GuildOverview.fxml
@@ -7,7 +7,6 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ButtonBar?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -41,30 +40,6 @@
                         <Font size="21.0" />
                      </font>
                   </Label>
-                  <Pane>
-                     <children>
-                        <VBox layoutX="6.0" layoutY="41.0" spacing="15.0">
-                           <children>
-                              <Label id="h1" text="Nyt Laug:">
-                                 <font>
-                                    <Font size="21.0" />
-                                 </font>
-                              </Label>
-                              <Label text="Laugets Tovholder:">
-                                 <font>
-                                    <Font size="21.0" />
-                                 </font>
-                              </Label>
-                           </children>
-                        </VBox>
-                        <VBox layoutX="191.0" layoutY="39.0" spacing="15.0">
-                           <children>
-                            </children>
-                        </VBox>
-                     </children>
-                  </Pane>
-                  
-                              <JFXTextField fx:id="txtGMCandidateSearch" onKeyTyped="#handleSearchForGM" promptText="Søg efter person" />
                            
                </children>
             </VBox>
@@ -85,8 +60,6 @@
             <Insets bottom="10.0" />
          </padding>
       </Pane>
-      <Pane />
-      <VBox fx:id="vBoxGuildOptions" layoutX="46.0" layoutY="414.0" spacing="15.0" stylesheets="@../../css/Style.css" />
       <ButtonBar layoutX="207.0" layoutY="447.0" nodeOrientation="RIGHT_TO_LEFT" prefHeight="40.0" AnchorPane.bottomAnchor="14.0" AnchorPane.leftAnchor="450.0">
         <buttons>
             <Button mnemonicParsing="false" onAction="#handleArchiveGuid" prefHeight="22.0" prefWidth="120.0" text="Arkiver Laug" />
@@ -98,7 +71,5 @@
             <Font size="21.0" />
          </font>
       </JFXTextArea>
-      <Button fx:id="btnAssignGM" layoutX="46.0" layoutY="245.0" mnemonicParsing="false" text="Button" />
-      <ListView fx:id="listPeople" layoutX="145.0" layoutY="176.0" onMouseClicked="#handleSelectGuildManager" prefHeight="200.0" prefWidth="200.0" />
    </children>
 </AnchorPane>
