+++ src/museumtimetracking/gui/views/root/activeGuilds/editGuild/EditGuildView.fxml
@@ -4,15 +4,33 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ButtonBar?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ListView?>
 <?import javafx.scene.control.TextArea?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<BorderPane fx:id="editGuildBorderPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="206.0" prefWidth="399.0" styleClass="borderPane" stylesheets="@../../../css/Style.css" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="museumtimetracking.gui.views.root.activeGuilds.editGuild.EditGuildViewController">
+<BorderPane fx:id="editGuildBorderPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" styleClass="borderPane" stylesheets="@../../../css/Style.css" xmlns="http://javafx.com/javafx/8.0.60" xmlns:fx="http://javafx.com/fxml/1" fx:controller="museumtimetracking.gui.views.root.activeGuilds.editGuild.EditGuildViewController">
    <center>
+      <Pane prefHeight="138.0" prefWidth="468.0" BorderPane.alignment="CENTER">
+         <children>
+            <ListView fx:id="listPeople" onMouseClicked="#handleSelectGuildManager" prefHeight="144.0" prefWidth="200.0" />
+            <Button fx:id="btnAssignGM" layoutX="389.0" layoutY="93.0" mnemonicParsing="false" onAction="#handleAssignGM" text="Button" />
+            <TextField fx:id="txtGMCandidateSearch" layoutX="218.0" layoutY="38.0" onKeyTyped="#handleSearchForGM" />
+         </children>
+      </Pane>
+   </center>
+   <bottom>
+      <ButtonBar nodeOrientation="RIGHT_TO_LEFT" BorderPane.alignment="CENTER">
+        <buttons>
+            <Button mnemonicParsing="false" onAction="#handleBack" text="Tilbage" />
+          <Button fx:id="btnSave" mnemonicParsing="false" onAction="#handleSaveGuild" text="Rediger" />
+        </buttons>
+      </ButtonBar>
+   </bottom>
+   <top>
       <VBox prefHeight="265.0" prefWidth="468.0" BorderPane.alignment="CENTER">
          <children>
             <Label text="Laug Navn">
@@ -33,18 +51,9 @@
             <TextArea fx:id="txtGuildDescription" prefHeight="200.0" prefWidth="200.0">
                <VBox.margin>
                   <Insets left="5.0" right="5.0" />
-               </VBox.margin></TextArea>
-            <HBox>
-               <children>
-                  <ButtonBar>
-                    <buttons>
-                      <Button fx:id="btnSave" mnemonicParsing="false" onAction="#handleSaveGuild" text="Rediger" />
-                        <Button mnemonicParsing="false" onAction="#handleBack" text="Tilbage" />
-                    </buttons>
-                  </ButtonBar>
-               </children>
-            </HBox>
+               </VBox.margin>
+            </TextArea>
          </children>
       </VBox>
-   </center>
+   </top>
 </BorderPane>
